requirements.txt
torch>=2.6.0
gymnasium>=1.0.0
numpy>=2.0.2
PyYAML>=6.0.2
opencv-python>=4.5.0
minigrid
imageio==2.4.0

